requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.6.0
anyio==4.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
config==0.5.1
contourpy==1.3.2
Corpus==0.4.2
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==6.4.2
email_validator==2.2.0
entrypoints==0.4
executing==2.1.0
Faker==30.8.1
fastapi==0.115.12
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
htbuilder==0.9.0
httpcore==1.0.2
httpx==0.28.1
httpx-sse==0.4.0
idna==3.7
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.31.0
nest-asyncio==1.6.0
nltk==3.9.1
numexpr==2.10.1
numpy==2.2.5
openai==1.77.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
pip==25.1
platformdirs==4.3.6
playsound==1.2.2
plotly==6.1.2
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==5.29.3
psutil==6.1.1
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.9.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.15
Pympler==1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.22.3
s3transfer==0.13.0
semver==3.0.4
setuptools==72.1.0
six==1.17.0
smmap==4.0.0
sniffio==1.3.0
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.39
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.2.2
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.32.1
validators==0.35.0
watchdog==4.0.2
wcwidth==0.2.13
wheel==0.45.1
win-inet-pton==1.1.0
yarl==1.20.0
youtube-search==2.1.2
zipp==3.22.0
zstandard==0.23.0
